CLI stuff should be in the cmd/ path
